requirements.txt
asgiref==3.5.0
cv==1.0.0
Django==4.0.3
django-scheduler==0.9.5
gunicorn==20.1.0
icalendar==4.0.9
MouseInfo==0.1.3
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.9
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.1
schedule==1.1.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
